mas-frontend/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
websockets==12.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0

# Google Cloud dependencies (matching MAS system)
google-auth==2.25.2
google-cloud-aiplatform==1.44.0
google-cloud-storage==2.13.0

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0

# Testing dependencies
aiohttp==3.9.1
websockets==12.0
colorama==0.4.6